Moved Config back to /src
